Apply hrm Feature #849851
Freeciv.org commit id: 3b69933736b8e8d505034ddcb1301e0200c1a582
